chore(license): consolidate keygen+sign CLIs into single tool
Reconcile duplicate license tooling per chat-time audit (PRs #21 + #29 surfaced
the duplication; this PR resolves it):

- port `--generate-keypair` from `generate-license-key.ts` into `issue-license.ts`
- delete `generate-license-key.ts` (functionality preserved + extended)
- update `license-crypto.ts` docstring to point at canonical script + paths
- realign `docs/PRODUCTION_LAUNCH_PLAN.md` H2/H3 to `docs/KEY_GENERATION.md`
  (the operator-facing runbook is the source of truth)
- simplify `docs/KEY_GENERATION.md` §2: replace the manual openssl + temp-file
  juggle with a single `--generate-keypair` invocation. No plaintext key files
  ever land on disk.

Canonical paths the entire toolchain now agrees on:

- revvault: `revdev/license-signing-private-key`, `revdev/license-signing-public-key`
- daemon env: `REVDEV_LICENSE_PUBLIC_KEY` (used by `verifyLicenseV2`)
- customer env: `REVEALUI_LICENSE_KEY` (the issued v2 license string)

Why `issue-license.ts` was kept (and `generate-license-key.ts` deleted):

- `issue-license.ts`'s revvault path matches the runbook the founder presumably
  tested. Switching to the other path means re-keying any existing local
  revvault entries.
- Better UX: `--help`, `--customer`, env/file fallback chain (env → revvault →
  `~/.revealui/license-private.pem`).
- The keygen mode was the only thing `generate-license-key.ts` did better, and
  it's ~30 LOC to port.

Verification:

- `scripts/issue-license.ts` transpiles strict-clean (tsc transpileModule).
- Pre-existing daemon typecheck failures (`zod`, `@neondatabase/serverless`,
  `@revealui/utils/logger`, `@revealui/core/observability`) confirmed unrelated
  to this PR via stash-and-retest on the same branch base.
- Did NOT live-test `--generate-keypair` (would write to revvault). The keygen
  block is a literal port of the well-tested logic from the deleted script;
  signing logic in this script is unchanged.
